requirements.txt
cryptography>=40.0.0
intelhex
click
cbor2
setuptools
pyyaml
pytest
littlefs-python
